requirements.txt
appnope==0.1.0
configparser==3.5.0
decorator==4.1.2
Django==2.0
django-extensions==1.9.8
djangorestframework==3.7.3
djangorestframework-gis==0.12
future==0.16.0
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
json-rpc==1.10.8
mccabe==0.6.1
mypy==0.550
numpy==1.13.3
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
psutil==5.4.2
psycopg2==2.7.3.2
ptyprocess==0.5.2
pycodestyle==2.3.1
pydocstyle==2.1.1
pyflakes==1.6.0
Pygments==2.2.0
python-language-server==0.11.1
pytz==2017.3
rope==0.10.7
Shapely==1.6.3
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
traitlets==4.3.2
typed-ast==1.1.0
typing==3.6.2
wcwidth==0.1.7
yapf==0.20.0
